requirements.txt
attrs==21.4.0
autobahn==19.11.2
Automat==20.2.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.11
constantly==15.1.0
cryptography==36.0.1
enum34==1.1.10
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
kiteconnect==4.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyOpenSSL==22.0.0
python-dateutil==2.8.2
python-dotenv==0.19.2
requests==2.27.1
service-identity==21.1.0
six==1.16.0
Twisted==21.7.0
txaio==21.2.1
typing_extensions==4.0.1
urllib3==1.26.8
zope.interface==5.4.0
